Make delay be a float not an int :)

+++ Assets/Scripts/UI/Sound/ProjectileSpawnerS/LongSoundPlayerInitialiser.cs
@@ -4,7 +4,7 @@ namespace BattleCruisers.UI.Sound.ProjectileSpawners
 {
     public class LongSoundPlayerInitialiser : ProjectileSoundPlayerInitialiser
     {
-        public int burstEndDelayInS;
+        public float burstEndDelayInS;
 
         protected override IProjectileSpawnerSoundPlayer CreateSoundPlayer(ISoundPlayerFactory soundPlayerFactory, ISoundKey firingSound, int burstSize, IAudioSourceWrapper audioSource)
         {
